# rest-app/dist/swagger.yaml
swagger: '2.0'
info:
  version: 0.1.0
  title: "hackzurich 2016 - AXA Winterthur - API"
  description: |
    AXA Winterthur is Gold sponsor "hackzurich" http://hackzurich.ch in 2016.
    The topic of AXA is "insurance as simple as a Game" aka "Clash of claims"
    the current API represents the most used domain objects of an insurance. we provide 2 different types: static / configation data and dynamic data, based on 10'000 predefined customers.
    ***
    dynamic data (prefilled with examples)
    - **customer data**: most attributes of a customer eg naming, address, date of birth, ...
    - **profile data**: important profile data of the customer
    - **contracts**: insurance contracts of the customer
    - **trips**: telmatics data of some trips, locations and events
    - **transactions**: financial transactions issued to 
    ***
    static / configuration data
    - **car + truck data**: some vehicle types and attributes, incl CO2 and energy attributes
    - **insurance types**: list of all types of insurances, group by L&S, P&C and health
    - **risk types**: risks possible to be able taken for the customer
    - **favorite risk object categories**: object categories to describe your favorite object
    ***
#host: hackzurich16.azurewebsites.net
host: localhost:3000

basePath: /axa
definitions:

  Trips:
    type: array
    items:
      $ref: '#/definitions/Trip'

  Trip:
    properties:
      id:
        type: string
        format: integer
        description: id of trip
      customer:
        type: string
        format: integer
        description: id of customer

  Transactions:
    type: array
    items:
      $ref: '#/definitions/Transaction'

  Transaction:
    properties:
      _id:
        type: string
        format: integer
        description: id of transaction
      customer:
        type: string
        format: integer
        description: id of customer
      amount:
        type: string
        format: double
        description: amount of transaction in EUR
      date:
        type: string
        format: date
        description: timestamp of transaction
      lineOfBusiness:
        type: string
        format: string
        description: insurance line of business where payment took place
        enum: 
         - P&C
         - L&C
         - Health
      name:
        type: string
        format: string
        description: name of receiver, receiver
      type:
        type: string
        format: string
        description: type of transaction.
        enum: 
         - deposit
         - invoice
         - payment
         - withdrawal
      account:
        type: string
        format: string
        description: account no of receiver


  CustomerResultList:
    type: object
    properties:
      data:
        $ref: '#/definitions/Customers'
      links:
        $ref: '#/definitions/Link'

  Customers:
    type: array
    items:
      $ref: '#/definitions/Customer'

  Customer:
    properties:
      id:
        type: string
        format: integer
        description: id of customer
      gender:
        type: string
        description: gender
        enum:
          - male
          - female
      title:
        type: string
        description: additional title of name e.g Dr.
      givenName:
        type: string
        description: first name 
      middleInitial:
        type: string
        description: short initial of middle name
      surname:
        type: string
        description: last name
      streetAddress:
        type: string
        description: address field of location incl number
      city:
        type: string
        description: city of address
      zipCode:
        type: integer
        description: zipcode of city
      country:
        type: string
        description: country ISO code
      countryFull:
        type: string
        description: full description of country based on ISO country code
      emailAddress:
        type: string
        description: primary email address
      username:
        type: string
        description: primary user name
      password:
        type: string
        description: visibile password. Just for testing purposes
      telephoneNumber:
        type: string
        description: primary phone number in local format. including prefix 0
      telephoneCountryCode:
        type: string
        description: country code of primary phone number without leading 00 or +
      birthday:
        type: string
        description: date of birth in form yyyy-mm-dd
      age:
        type: integer
        description: calculated date based on year 2016. only for tests - not dynamic
      occupation:
        type: string
        description: endlish expression of job title
      company:
        type: string
        description: name of company
      vehicle:
        type: string
        description: name of car vehicle type
      kilograms:
        type: integer
        description: current weight of customer at 16.9.2016, hackzurich
      centimeters:
        type: integer
        description: body height in cm
      location:
        $ref: '#/definitions/Location2Dsphere'
            
  Location2Dsphere:
    title: Spherical 2D location
    description: Location structure of 2D spherical point positioning with long and lat
    properties:
      type:
        description: type of 2D representation. here always point
        default: 
          "Point"
        type: string
        enum:
          - Point
      coordinate:
        description: array of 2 integer values with longitude and latitude
        type: array
        items:
          type: number
          format: double

  Link:
    title: navigation links of result lists
    properties:
      cur:
        type: string
        description: URL of the current page request. including needed skip + limit parameters
      first:
        type: string
        description: URL for first paging call
      prev:
        type: string
        description: URL for prev paging call. Can be null if no prev possible, because already on first page
      next:
        type: string
        description: URL for next paging call. Can be null if no next possible, because already on last page
      last:
        type: string
        description: URL for last paging call
      count:
        type: number
        format: integer
        description: nof of results in this page. Always <= limit
      totalCount:
        type: number
        format: integer
        description: total count of results for query
paths:
  /customers:
    get:
      summary: list of all customers, use ranges
      description: |
        get all customers using. max limit 100, using skip + limit to range
      tags:
        - Customer
      parameters:
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer

      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/CustomerResultList'
            
  /customers/search/byZip/{zip}:
    get:
      summary: search customers by zip code, use ranges
      description: |
        search all customers at their location by zip, using. max limit 100, using skip + limit to range
      parameters:
        - name: zip
          in: path
          description: zip code of location
          required: true
          type: integer
          format: integer
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: invalid zip parameter
        '500':
          description: technical server error

  /customers/search/byName/{name}:
    get:
      summary: search customers by lastname or firstname, using *wildcards*, use ranges
      description: |
        search all customers by their lastname or firstname using wildcards automatically, using. max limit 100, using skip + limit to range
      parameters:
        - name: name
          in: path
          description: part of name (first or last)
          required: true
          type: string
          format: string
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: invalid name parameter
        '500':
          description: technical server error

  /customers/search/byWord/{text}:
    get:
      summary: search customers by any word in all indexed fields, use ranges
      description: |
        search all customers by any word in all indexed fields using full text search. Splitted words with blanks used as OR operator. using max limit 100, using skip + limit to range
      parameters:
        - name: text
          in: path
          description: word to search
          required: true
          type: string
          format: string
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: invalid text parameter
        '500':
          description: technical server error
          
  /customers/search/byQuery/{query}/{sort}:
    get:
      summary: search customers using mongodb query, use ranges
      description: |
        search all customers using a mongodb query in JSON format - see http://bit.ly/2aI9Djk, using. max limit 100, using skip + limit to range
      parameters:
        - name: query
          in: path
          description: mongodb query in JSON format - see http://bit.ly/2aI9Djk
          required: true
          default: "{}"
          type: string
          format: string
        - name: sort
          in: path
          description: sort column in JSON format, use id as default - see http://bit.ly/2aIaK2I
          required: true
          default: '{"id" : 1}'
          type: string
          format: string
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: invalid query parameter
        '500':
          description: technical server error

  /customers/search/near/{longitude},{latitude},{meter}:
    get:
      summary: search nearest customers in meter to a given location
      description: |
        search nearest customers in meter to a given location. Max limit 100, using skip + limit to range
      parameters:
        - name: longitude 
          in: path
          description: longitude of given location to search
          required: true
          type: number
          default: 46.731947
          format: double
        - name: latitude
          in: path
          description: latitude of given location to search
          required: true
          type: number
          default: 6.962926
          format: double
        - name: meter
          in: path
          description: define nearest area to location in meter
          required: true
          default: 1000
          type: number
          format: integer
        - name: skip
          in: query
          description: nof items skipped in result. Default = 0. Used with limit for paging
          required: false
          type: integer
          format: integer
        - name: limit
          in: query
          description: limit no of responses. Default = 10. Used with skip for paging
          required: false
          type: integer
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: invalid query parameter
        '500':
          description: technical server error


  /customers/{id}:
    get:
      summary: get customer with id
      description: |
        get all attributes of a customer based on id
      parameters:
        - name: id
          in: path
          description: id of customer
          required: true
          type: string
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Customer'
        '404':
          description: customer not found
        '500':
          description: technical server error
          
  /customers/{id}/trips:
    get:
      summary: get all telematic trips and events
      description: |
        get all telematics events and positioning information of the customers trips
      parameters:
        - name: id
          in: path
          description: id of customer
          required: true
          type: string
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Trips'
        '404':
          description: customer not found
        '500':
          description: technical server error
          
  /customers/{id}/transactions:
    get:
      summary: get some financial transactions of this customer
      description: |
        get all financial transactions of this customer
      parameters:
        - name: id
          in: path
          description: id of customer
          required: true
          type: string
          format: integer
      tags:
        - Customer
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Transactions'
        '404':
          description: customer not found
        '500':
          description: technical server error